+++ R/google_map_layer_fusion.R
@@ -30,6 +30,14 @@
 #' google_map(key = map_key, location = c(41.8, -87.7), zoom = 9) %>%
 #'   add_fusion(query = qry)
 #'
+#'
+#' qry <- list(select = 'address',
+#'     from = '1d7qpn60tAvG4LEg4jvClZbc1ggp8fIGGvpMGzA',
+#'     where = 'ridership > 200')
+#'
+#' google_map(key = map_key, location = c(41.8, -87.7), zoom = 9) %>%
+#'   add_fusion(query = qry)
+#'
 #' qry <- data.frame(select = 'geometry',
 #'    from = '1ertEwm-1bMBhpEwHhtNYT47HQ9k2ki_6sRa-UQ')
 #'
@@ -55,14 +63,14 @@
 #' styles <- '[
 #'   {
 #'     "polygonOptions":{
-#'       "fillColor":"#00FF00",
+#'       "fillColor":"#FFFF00",
 #'       "fillOpacity":0.3
 #'       }
 #'     },
 #'     {
 #'       "where":"birds > 300",
 #'        "polygonOptions":{
-#'          "fillColor":"#0000FF"
+#'          "fillColor":"#000000"
 #'        }
 #'      },
 #'      {
@@ -72,6 +80,12 @@
 #'          }
 #'      }
 #'    ]'
+#'
+#' google_map(key = map_key, location = c(-25.3, 133), zoom = 4) %>%
+#'   add_fusion(query = qry, styles = styles)
+#'
+#' ## using a JSON style
+#' attr(styles, 'class') <- 'json'
 #'
 #' google_map(key = map_key, location = c(-25.3, 133), zoom = 4) %>%
 #'   add_fusion(query = qry, styles = styles)
@@ -90,12 +104,13 @@ add_fusion <- function(map, query, styles = NULL, heatmap = FALSE, layer_id = NU
   ## TODO:
   ## - update bounds on layer
   ## - info window
-  ## - allow JSON style
 
   logicalCheck(heatmap)
   layer_id <- layerId(layer_id)
 
   query <- validateFusionQuery(query)
+
+  print(query)
 
   if(!is.null(styles))
     styles <- validateFusionStyle(styles)